move symmetry file to specs
